Packages in domain refactor, use cases
