+++ src/main/java/com/alkemy/ong/domain/activity/Activity.java
@@ -1,4 +1,4 @@
-package com.alkemy.ong.domain.models;
+package com.alkemy.ong.domain.activity;
 
 public class Activity {
   String name;
